requirements-dev.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
dj-database-url==0.5.0
dj-static==0.0.6
python-decouple==3.3